Added many files that were used as baseline.  The fortran mars agrees with the baseline except for the timing which is understanable.  pymars does NOT agree with the fortran which is troubling.  I'm investigating if I have the wrong version of pymars, This requires a new branch that I'll call pymars.
I also fixed a problem with a parameter named txl that in some situations was being
returned before assignmanet. This was changed marsgo.py. Removing txl from the
parameters returned appeared to be benign but I'm not sure.  I never had the
warning before of a reference before an assignment. That suggests that I don't
have the right version. I happen to have a disk from 06_17_2013 that I believe
is a reliable version. Stay tuned.